SmartLead Daily Automation/Automation/requirements.txt
requests
python-dotenv
sendgrid
urllib3
